Added map, UI, shop, turret placement, started upgrades
Best practice would be to intricately detail changes here/break this up
into multiple commits. However, I have been working on this non stop since
this morning and my brain is mush. I'm going to sleep now. Any issues will
be dealt with in the morning, but all looks good as of this commit.